refactor(naming): newCommunity -> createCommunity
