+++ server/routes.js
@@ -93,8 +93,8 @@ router.route('/:community/repo/:id').get((req, res) => {
 })
 
 // 创建新社区
-router.route('/discovery/new').get((req, res) => {
-  return renderAndCache({ req, res })
+router.route('/create/community').get((req, res) => {
+  return renderAndCache({ req, res, page: '/create/community' })
 })
 // 所有社区
 router.route('/discovery').get((req, res) => res.redirect('/discovery/pl'))
