+++ src/containers/content/CreateCommunityContent/Banner/InputBox.js
