+++ src/containers/content/CreateCommunityContent/Banner/SelectType/TypeBoxes.js
