+++ src/containers/content/CreateCommunityContent/Banner/SelectType/index.js
