+++ src/containers/content/CreateCommunityContent/Banner/SetupInfo/index.js
