+++ src/containers/content/CreateCommunityContent/Banner/index.js
