+++ src/containers/content/CreateCommunityContent/Content/SelectType.js
