+++ src/containers/content/CreateCommunityContent/SlogenText.js
