+++ src/containers/content/CreateCommunityContent/index.js
@@ -17,7 +17,7 @@ import { useInit } from './logic'
 /* eslint-disable-next-line */
 const log = buildLog('C:DiscoveryContent')
 
-const NewCommunityContentContainer = ({ newCommunityContent: store }) => {
+const CreateCommunityContentContainer = ({ createCommunityContent: store }) => {
   useInit(store)
   const { step, selectTypeStatus, setupDomainStatus, setupInfoStatus } = store
 
@@ -41,8 +41,8 @@ const NewCommunityContentContainer = ({ newCommunityContent: store }) => {
   )
 }
 
-NewCommunityContentContainer.getInitialProps = async () => ({
+CreateCommunityContentContainer.getInitialProps = async () => ({
   namespacesRequired: ['common'],
 })
 
-export default connectStore(NewCommunityContentContainer)
+export default connectStore(CreateCommunityContentContainer)
