+++ src/containers/content/CreateCommunityContent/styles/banner/index.js
