+++ src/containers/content/CreateCommunityContent/styles/banner/input_box.js
