+++ src/containers/content/CreateCommunityContent/styles/banner/select_type/index.js
