+++ src/containers/content/CreateCommunityContent/styles/banner/select_type/type_boxes.js
