+++ src/containers/content/CreateCommunityContent/styles/banner/setup_domain/index.js
