+++ src/containers/content/CreateCommunityContent/styles/content/demo_community.js
