+++ src/containers/content/CreateCommunityContent/styles/content/fake_browser/content.js
