+++ src/containers/content/CreateCommunityContent/styles/content/fake_browser/index.js
