+++ src/containers/content/CreateCommunityContent/styles/content/index.js
