+++ src/containers/content/CreateCommunityContent/styles/content/select_type.js
