+++ src/containers/content/CreateCommunityContent/styles/index.js
